Rename sourced_models.py to models.py for cleaner naming
- Simplified module naming from sourced_models to models
- Updated import in mcp_server.py accordingly
- Maintains all functionality while improving clarity
